requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# AI & Content Generation
# anthropic==0.7.7  # Claude API (alternative)
openai==1.54.0  # OpenAI API (GPT-4)

# Email
resend==0.6.0

# Task Queue (Future: Celery + Redis)
# celery==5.3.4
# redis==5.0.1

# Utilities
loguru==0.7.2
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
email-validator==2.1.0
httpx==0.25.2
